move wait methods around a bit
