+++ lib/rsmp.rb
@@ -20,6 +20,7 @@ require 'rsmp/components'
 require 'rsmp/site'
 require 'rsmp/proxy'
 require 'rsmp/supervisor_proxy'
+require 'rsmp/site_proxy_wait'
 require 'rsmp/site_proxy'
 require 'rsmp/error'
 require 'rsmp/probe'
